feat(tui): implement Milestone 4 - Loop Runner Dashboard MVP
Add Run Dashboard screen with real-time loop execution monitoring:

- Multi-pane layout: status header, output viewer, criteria sidebar,
  events log, and git status pane
- Background async loop runner with event-based TUI updates
- Cancellation support via Ctrl+C or Esc during active runs
- Scrollable output with follow mode toggle (f key)
- Parse and display completion criteria from PROMPT.md

Performance improvements from Gemini review:
- Convert blocking std::fs to async tokio::fs in runner
- Move git operations to background thread via spawn_blocking
- Remove 500-char output truncation (TUI handles scrolling)
- Use actual cooldown remaining time instead of fixed 5 seconds
- Change iteration type from u32 to u64 to prevent overflow

+++ crates/ralf-engine/src/changelog.rs
@@ -13,7 +13,7 @@ pub struct ChangelogEntry<'a> {
     /// Run identifier.
     pub run_id: &'a str,
     /// Iteration number.
-    pub iteration: u32,
+    pub iteration: u64,
     /// Model invocation result.
     pub invocation: &'a InvocationResult,
     /// Verifier results.
